+++ kroto-plus-coroutines/src/main/kotlin/com/github/marcoferrer/krotoplus/coroutines/RpcBidiChannel.kt
@@ -6,18 +6,23 @@ import kotlinx.coroutines.channels.SendChannel
 
 
 @Deprecated(
-    "RpcBidiChannel has been renamed to ClientBidiCallChannel",
-    ReplaceWith("ClientBidiCallChannel<ReqT,RespT>")
+    message = "Deprecated in favor of back-pressure supporting APIs.",
+    level = DeprecationLevel.WARNING,
+    replaceWith = ReplaceWith("ClientBidiCallChannel<ReqT,RespT>")
 )
 typealias RpcBidiChannel<ReqT, RespT> = ClientBidiCallChannel<ReqT, RespT>
 
 @Deprecated(
-    "ClientBidiCallChannel has been moved",
-    ReplaceWith("com.github.marcoferrer.krotoplus.coroutines.client.ClientBidiCallChannel<ReqT,RespT>")
+    message = "Deprecated in favor of back-pressure supporting APIs.",
+    level = DeprecationLevel.WARNING,
+    replaceWith = ReplaceWith("com.github.marcoferrer.krotoplus.coroutines.client.ClientBidiCallChannel<ReqT,RespT>")
 )
 typealias ClientBidiCallChannel<ReqT, RespT> = com.github.marcoferrer.krotoplus.coroutines.client.ClientBidiCallChannel<ReqT, RespT>
 
-@Deprecated("Deprecated in favor of using kroto generated base impl.")
+@Deprecated(
+    message = "Deprecated in favor of back-pressure supporting APIs.",
+    level = DeprecationLevel.WARNING
+)
 data class ServerBidiCallChannel<ReqT, RespT>(
     val requestChannel: ReceiveChannel<ReqT>,
     val responseChannel: SendChannel<RespT>