+++ kroto-plus-coroutines/src/main/kotlin/com/github/marcoferrer/krotoplus/coroutines/SuspendingUnaryObserver.kt
@@ -5,6 +5,7 @@ import kotlin.coroutines.Continuation
 import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
 
+@KrotoPlusInternalApi
 class SuspendingUnaryObserver<RespT>(
     @Volatile @JvmField private var cont: Continuation<RespT>?
 ) : StreamObserver<RespT> {