+++ kroto-plus-coroutines/src/main/kotlin/com/github/marcoferrer/krotoplus/coroutines/server/ServerChannels.kt
@@ -1,10 +1,9 @@
 package com.github.marcoferrer.krotoplus.coroutines.server
 
-import com.github.marcoferrer.krotoplus.coroutines.FlowControlledObserver
+import com.github.marcoferrer.krotoplus.coroutines.call.FlowControlledObserver
 import io.grpc.stub.CallStreamObserver
 import io.grpc.stub.StreamObserver
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.channels.ReceiveChannel
 import java.util.concurrent.atomic.AtomicBoolean