+++ protoc-gen-kroto-plus/src/main/kotlin/com/github/marcoferrer/krotoplus/generators/GrpcStubExtsGenerator.kt
@@ -1,14 +1,14 @@
 package com.github.marcoferrer.krotoplus.generators
 
 import com.github.marcoferrer.krotoplus.config.GrpcStubExtsGenOptions
-import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
 import com.github.marcoferrer.krotoplus.generators.Generator.Companion.AutoGenerationDisclaimer
 import com.github.marcoferrer.krotoplus.proto.ProtoMethod
 import com.github.marcoferrer.krotoplus.proto.ProtoService
-import com.github.marcoferrer.krotoplus.utils.CommonClassNames
-import com.github.marcoferrer.krotoplus.utils.matches
+import com.github.marcoferrer.krotoplus.utils.*
 import com.google.protobuf.compiler.PluginProtos
 import com.squareup.kotlinpoet.*
+import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
+import io.grpc.MethodDescriptor.*
 
 
 object GrpcStubExtsGenerator : Generator {
@@ -19,11 +19,13 @@ object GrpcStubExtsGenerator : Generator {
     override fun invoke(): PluginProtos.CodeGeneratorResponse {
         val responseBuilder = PluginProtos.CodeGeneratorResponse.newBuilder()
 
-        for (service in context.schema.protoServices) {
-            for (options in context.config.grpcStubExtsList) {
+        for (options in context.config.grpcStubExtsList) {
+            val fileBuilder = FileBuilder(options)
 
-                if (options.filter.matches(service.protoFile.name)) {
-                    buildFileSpec(service, options)?.let { responseBuilder.addFile(it) }
+            for (service in context.schema.protoServices) {
+
+                fileBuilder.buildFile(service)?.let { file ->
+                    responseBuilder.addFile(file)
                 }
             }
         }
@@ -31,175 +33,192 @@ object GrpcStubExtsGenerator : Generator {
         return responseBuilder.build()
     }
 
-    private fun buildFileSpec(
-        service: ProtoService,
-        options: GrpcStubExtsGenOptions
-    ): PluginProtos.CodeGeneratorResponse.File? {
-        val filename = "${service.name}RpcOverloads"
-        val fileSpecBuilder = FileSpec
-            .builder(service.protoFile.javaPackage, filename)
-            .addComment(AutoGenerationDisclaimer)
-            .addAnnotation(
-                AnnotationSpec.builder(JvmName::class.asClassName())
-                    .useSiteTarget(AnnotationSpec.UseSiteTarget.FILE)
-                    .addMember("%S", "-$filename")
-                    .build()
-            )
+    class FileBuilder(val options: GrpcStubExtsGenOptions) {
 
-        for (method in service.methodDefinitions)
-            buildFunSpecs(method, options, fileSpecBuilder)
+        fun buildFile(service: ProtoService): PluginProtos.CodeGeneratorResponse.File? = with(service) {
+            if (!options.filter.matches(protoFile.name))
+                return null
 
-        return fileSpecBuilder.build()
-            .takeIf { it.members.isNotEmpty() }
-            ?.toResponseFileProto()
-    }
+            val filename = "${name}RpcOverloads"
+            val fileSpecBuilder = FileSpec
+                .builder(protoFile.javaPackage, filename)
+                .addComment(AutoGenerationDisclaimer)
+                .addAnnotation(
+                    AnnotationSpec.builder(JvmName::class.asClassName())
+                        .useSiteTarget(AnnotationSpec.UseSiteTarget.FILE)
+                        .addMember("%S", "-$filename")
+                        .build()
+                )
 
-    private fun buildFunSpecs(
-        method: ProtoMethod,
-        options: GrpcStubExtsGenOptions,
-        fileSpecBuilder: FileSpec.Builder
-    ) {
-        when {
-            //Unary
-            method.isUnary ->
-                buildUnaryOverloads(method, options, fileSpecBuilder)
+            fileSpecBuilder.apply {
+                addFunctions(buildDefaultStubExts())
 
-            //Server Streaming //TODO Add support for Blocking Server Streaming
-            options.supportCoroutines && method.isServerStream ->
-                buildServerStreamingOverloads(method, fileSpecBuilder)
+                if (options.supportCoroutines) {
+                    addFunctions(buildStubCoroutineExts())
+                    addFunctions(buildClientStubRpcRequestOverloads())
+                }
+            }
 
-            //Bidi && Client Streaming
-            options.supportCoroutines && (method.isBidi || method.isClientStream) ->
-                buildBidiStreamingOverloads(method, fileSpecBuilder)
+            return fileSpecBuilder
+                .build()
+                .takeIf { it.members.isNotEmpty() }
+                ?.toResponseFileProto()
         }
-    }
 
-    private fun buildUnaryOverloads(
-        method: ProtoMethod,
-        options: GrpcStubExtsGenOptions,
-        fileSpecBuilder: FileSpec.Builder
-    ) {
+        private fun ProtoService.buildDefaultStubExts(): List<FunSpec> =
+            methodDefinitions
+                .filter { it.isUnary }
+                .flatMap { it.buildUnaryDefaultOverloads() }
 
-        val funSpecBuilder = FunSpec.builder(method.functionName)
+        private fun ProtoService.buildStubCoroutineExts(): List<FunSpec> =
+            methodDefinitions.map { method ->
+                when (method.type) {
+                    MethodType.UNARY -> method.buildStubUnaryMethod()
+                    MethodType.CLIENT_STREAMING -> method.buildStubClientStreamingMethod()
+                    MethodType.SERVER_STREAMING -> method.buildStubServerStreamingMethod()
+                    MethodType.BIDI_STREAMING -> method.buildStubBidiStreamingMethod()
+                    MethodType.UNKNOWN -> throw IllegalStateException("Unknown method type")
+                }
+            }
 
-        if (method.isNotEmptyInput) {
-            funSpecBuilder
+        private fun ProtoService.buildClientStubRpcRequestOverloads(): List<FunSpec> =
+            methodDefinitions.mapNotNull {
+                when {
+
+                    it.type == MethodType.UNARY && it.isNotEmptyInput ->
+                        it.buildUnaryCoroutineExtOverload()
+
+                    it.type == MethodType.SERVER_STREAMING ->
+                        it.buildStubServerStreamingMethodOverload()
+
+                    else -> null
+                }
+            }
+
+
+        private fun ProtoMethod.buildUnaryDefaultOverloads(): List<FunSpec> {
+
+            val funSpecs = mutableListOf<FunSpec>()
+
+            // Future Stub Ext
+            funSpecs += FunSpec.builder(functionName)
                 .addModifiers(KModifier.INLINE)
-                .addParameter(
-                    "block", LambdaTypeName.get(
-                        receiver = method.requestClassName.nestedClass("Builder"),
-                        returnType = UNIT
+                .addCode(requestClassName.requestValueBuilderCodeBlock)
+                .addStatement("return %N(request)", functionName)
+                .addParameter("block", requestClassName.builderLambdaTypeName)
+                .receiver(protoService.futureStubClassName)
+                .returns(CommonClassNames.listenableFuture.parameterizedBy(responseClassName))
+                .build()
+
+            // Future Stub NoArg Ext
+            funSpecs += FunSpec.builder(functionName)
+                .addStatement("return %N(%T.getDefaultInstance())", functionName, requestClassName)
+                .receiver(protoService.futureStubClassName)
+                .returns(CommonClassNames.listenableFuture.parameterizedBy(responseClassName))
+                .build()
+
+            // Blocking Stub Ext
+            funSpecs += FunSpec.builder(functionName)
+                .addModifiers(KModifier.INLINE)
+                .addCode(requestClassName.requestValueBuilderCodeBlock)
+                .addStatement("return %N(request)", functionName)
+                .addParameter("block", requestClassName.builderLambdaTypeName)
+                .receiver(protoService.blockingStubClassName)
+                .returns(responseClassName)
+                .build()
+
+            // Blocking Stub NoArg Ext
+            funSpecs += FunSpec.builder(functionName)
+                .addStatement("return %N(%T.getDefaultInstance())", functionName, requestClassName)
+                .receiver(protoService.blockingStubClassName)
+                .returns(responseClassName)
+                .build()
+
+            return funSpecs
+        }
+
+        private fun ProtoMethod.buildUnaryCoroutineExtOverload(): FunSpec =
+            FunSpec.builder(functionName)
+                .addModifiers(KModifier.SUSPEND)
+                .receiver(protoService.asyncStubClassName)
+                .addModifiers(KModifier.INLINE)
+                .addParameter("block", requestClassName.builderLambdaTypeName)
+                .returns(responseClassName)
+                .addCode(requestClassName.requestValueBuilderCodeBlock)
+                .addStatement("return %N(request)", functionName)
+                .build()
+
+        private fun ProtoMethod.buildStubServerStreamingMethodOverload(): FunSpec =
+            FunSpec.builder(functionName)
+                .receiver(protoService.asyncStubClassName)
+                .returns(CommonClassNames.receiveChannel.parameterizedBy(responseClassName))
+                .addModifiers(KModifier.INLINE)
+                .addParameter("block", requestClassName.builderLambdaTypeName)
+                .addCode(requestClassName.requestValueBuilderCodeBlock)
+                .addStatement("return %N(request)", functionName)
+                .build()
+
+        private fun ProtoMethod.buildStubUnaryMethod(): FunSpec =
+            FunSpec.builder(functionName)
+                .addModifiers(KModifier.SUSPEND)
+                .returns(responseClassName)
+                .receiver(protoService.asyncStubClassName)
+                .addParameter(requestClassName.requestParamSpec)
+                .addStatement(
+                    "return %T(request, %T.%N())",
+                    CommonClassNames.ClientCalls.clientCallUnary,
+                    protoService.enclosingServiceClassName,
+                    methodDefinitionGetterName
+                )
+                .build()
+
+        private fun ProtoMethod.buildStubClientStreamingMethod(): FunSpec =
+            FunSpec.builder(functionName)
+                .receiver(protoService.asyncStubClassName)
+                .returns(
+                    CommonClassNames.ClientChannels.clientStreamingCallChannel.parameterizedBy(
+                        requestClassName,
+                        responseClassName
                     )
                 )
-        }
-
-        funSpecBuilder
-            .receiver(method.protoService.futureStubClassName)
-            .addCode(method.requestValueCodeBlock())
-            .addStatement("return %N(request)", method.functionName)
-            .returns(
-                ClassName("com.google.common.util.concurrent", "ListenableFuture")
-                    .parameterizedBy(method.responseClassName)
-            )
-            .also { fileSpecBuilder.addFunction(it.build()) }
-
-        funSpecBuilder
-            .receiver(method.protoService.blockingStubClassName)
-            .returns(method.responseClassName)
-            .also { fileSpecBuilder.addFunction(it.build()) }
-
-        if (options.supportCoroutines)
-            buildSuspendingUnaryOverloads(method, fileSpecBuilder)
-    }
-
-    private fun buildSuspendingUnaryOverloads(method: ProtoMethod, fileSpecBuilder: FileSpec.Builder) {
-
-        FunSpec.builder(method.functionName)
-            .addModifiers(KModifier.SUSPEND)
-            .receiver(method.protoService.asyncStubClassName)
-            .apply {
-                if (method.isNotEmptyInput)
-                    addParameter("request", method.requestClassName) else
-                    addCode(method.requestValueCodeBlock())
-            }
-            .returns(method.responseClassName)
-            .addStatement(
-                "return %T{ observer -> %N(request,observer) }",
-                CommonClassNames.suspendingUnaryCallObserver,
-                method.functionName
-            )
-            .also { fileSpecBuilder.addFunction(it.build()) }
-
-        if (method.isNotEmptyInput) {
-            FunSpec.builder(method.functionName)
-                .addModifiers(KModifier.SUSPEND)
-                .receiver(method.protoService.asyncStubClassName)
-                .apply {
-                    if (method.isNotEmptyInput) {
-                        addModifiers(KModifier.INLINE)
-                        addParameter(
-                            "block", LambdaTypeName.get(
-                                receiver = method.requestClassName.nestedClass("Builder"),
-                                returnType = UNIT
-                            )
-                        )
-                    }
-                }
-                .returns(method.responseClassName)
-                .addCode(method.requestValueCodeBlock())
-                .addStatement("return %N(request)", method.functionName)
-                .also { fileSpecBuilder.addFunction(it.build()) }
-        }
-    }
-
-    private fun buildServerStreamingOverloads(method: ProtoMethod, fileSpecBuilder: FileSpec.Builder) {
-
-        val returnType = CommonClassNames.inboundStreamChannel
-            .parameterizedBy(method.responseClassName)
-
-        FunSpec.builder(method.functionName)
-            .receiver(method.protoService.asyncStubClassName)
-            .apply {
-                if (method.isNotEmptyInput)
-                    addParameter("request", method.requestClassName) else
-                    addCode(method.requestValueCodeBlock())
-            }
-            .returns(returnType)
-            .addStatement(
-                "return %T().also { observer -> %N(request,observer) }",
-                returnType,
-                method.functionName
-            )
-            .also { fileSpecBuilder.addFunction(it.build()) }
-    }
-
-
-    private fun buildBidiStreamingOverloads(method: ProtoMethod, fileSpecBuilder: FileSpec.Builder) {
-
-        FunSpec.builder(method.functionName)
-            .receiver(method.protoService.asyncStubClassName)
-            .addAnnotation(CommonClassNames.obsoleteCoroutinesApi)
-            .addAnnotation(CommonClassNames.experimentalCoroutinesApi)
-            .addAnnotation(CommonClassNames.experimentalKrotoPlusCoroutinesApi)
-            .returns(
-                CommonClassNames.clientBidiCallChannel.parameterizedBy(
-                    method.requestClassName,
-                    method.responseClassName
+                .addStatement(
+                    "return %T(%T.%N())",
+                    CommonClassNames.ClientCalls.clientCallClientStreaming,
+                    protoService.enclosingServiceClassName,
+                    methodDefinitionGetterName
                 )
-            )
-            .addStatement(
-                "return %T{ responseObserver -> %N(responseObserver) }",
-                CommonClassNames.bidiCallChannel,
-                method.functionName
-            )
-            .also { fileSpecBuilder.addFunction(it.build()) }
-    }
+                .build()
 
-    private fun ProtoMethod.requestValueCodeBlock(): CodeBlock {
-        val requestValueTemplate = if (isEmptyInput)
-            "val request = %T.getDefaultInstance()\n" else
-            "val request = %T.newBuilder().apply(block).build()\n"
 
-        return CodeBlock.of(requestValueTemplate, requestClassName)
+        private fun ProtoMethod.buildStubServerStreamingMethod(): FunSpec =
+            FunSpec.builder(functionName)
+                .returns(CommonClassNames.receiveChannel.parameterizedBy(responseClassName))
+                .receiver(protoService.asyncStubClassName)
+                .addParameter(requestClassName.requestParamSpec)
+                .addStatement(
+                    "return %T(request, %T.%N())",
+                    CommonClassNames.ClientCalls.clientCallServerStreaming,
+                    protoService.enclosingServiceClassName,
+                    methodDefinitionGetterName
+                )
+                .build()
+
+        private fun ProtoMethod.buildStubBidiStreamingMethod(): FunSpec =
+            FunSpec.builder(functionName)
+                .receiver(protoService.asyncStubClassName)
+                .returns(
+                    CommonClassNames.ClientChannels.clientBidiCallChannel.parameterizedBy(
+                        requestClassName,
+                        responseClassName
+                    )
+                )
+                .addStatement(
+                    "return %T(%T.%N())",
+                    CommonClassNames.ClientCalls.clientCallBidiStreaming,
+                    protoService.enclosingServiceClassName,
+                    methodDefinitionGetterName
+                )
+                .build()
+
     }
 }